Mutarea functionalitate
Functionalitatea de a parsa fisiere JSON a fost mutata
in clasa Reader.
